+++ code/testing/languages/nl.f1re.testing/source_gen/nl/f1re/testing/structure/LanguageConceptSwitch.java
@@ -9,10 +9,12 @@ import org.jetbrains.mps.openapi.language.SAbstractConcept;
 
 public final class LanguageConceptSwitch {
   private final LanguageConceptIndex myIndex;
-  public static final int FileNodeEditorExpression = 0;
+  public static final int EditorTestLifecycleMethods = 0;
+  public static final int FileNodeEditorExpression = 1;
 
   public LanguageConceptSwitch() {
     LanguageConceptIndexBuilder builder = new LanguageConceptIndexBuilder(0x953e4089c643455bL, 0x8629636de7085d1cL);
+    builder.put(0x1c7d44aa128f6ec2L, EditorTestLifecycleMethods);
     builder.put(0x2923b546f4764a73L, FileNodeEditorExpression);
     myIndex = builder.seal();
   }